Make spectroscopy plan top level
